+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   BREAKING: rename Algo to Function (#170)
+-   Use exportPerformance endpoint with metadata for every download (#156)
 
 ### Fixed